Move UserDefaultsFeedStore to its own file in FeedStoreChallenge Group
